move and rename subpackage
