+++ handler/receiver_test.go
@@ -1,4 +1,4 @@
-package receiver
+package handler
 
 import (
 	"encoding/json"
